update: marker, layers, auth and etc

+++ src/modules/layer/layer.controller.ts
@@ -6,24 +6,22 @@ import {
   Param,
   Patch,
   Post,
-  Req,
   UseGuards,
 } from '@nestjs/common';
-import { Request } from 'express';
-
 import { LayerService } from './layer.service';
 import { CreateLayerDto } from './dto/create-layer.dto';
 import { UpdateLayerDto } from './dto/update-layer.dto';
-import { AuthGuard } from '../auth/guard/auth.guard';
+import { JwtAuthGuard } from '@/modules/auth/guard/jwt-auth.guard';
+import { CurrentUser } from '@/decorators/current-user.decorator';
+import { User } from '@/modules/user/entities/user.entity';
 
 @Controller('layers')
 export class LayerController {
   constructor(private layerService: LayerService) {}
 
   @Get()
-  getAll(@Req() req: Request) {
-    const { refresh_token } = req.cookies;
-    return this.layerService.getAll(refresh_token);
+  getAll() {
+    return this.layerService.getAll();
   }
 
   @Get('/:id')
@@ -31,22 +29,25 @@ export class LayerController {
     return this.layerService.getOne(id);
   }
 
-  @UseGuards(AuthGuard)
-  @Post('/create')
-  createOne(@Body() dto: CreateLayerDto, @Req() req: Request) {
-    const { refresh_token } = req.cookies;
-    return this.layerService.createOne(dto, refresh_token);
+  @UseGuards(JwtAuthGuard)
+  @Post()
+  createOne(@CurrentUser() user: User, @Body() dto: CreateLayerDto) {
+    return this.layerService.createOne(user.id, dto);
   }
 
-  @UseGuards(AuthGuard)
+  @UseGuards(JwtAuthGuard)
   @Patch('/:id')
-  updateOne(@Param() { id }, @Body() dto: UpdateLayerDto) {
-    return this.layerService.updateOne(id, dto);
+  updateOne(
+    @Param() { id },
+    @CurrentUser() user: User,
+    @Body() dto: UpdateLayerDto,
+  ) {
+    return this.layerService.updateOne(id, user.id, dto);
   }
 
-  @UseGuards(AuthGuard)
+  @UseGuards(JwtAuthGuard)
   @Delete('/:id')
-  deleteOne(@Param() { id }) {
-    return this.layerService.deleteOne(id);
+  deleteOne(@Param() { id }, @CurrentUser() user: User) {
+    return this.layerService.deleteOne(id, user.id);
   }
 }
